refactor off treeops
